+++ src/soliplex/secrets.py
@@ -1,10 +1,8 @@
+from __future__ import annotations  # forward refs in typing decls
+
 import os
 import pathlib
 import subprocess
-
-from soliplex.config import secrets as config_secrets
-
-_secrets_getters = config_secrets.SECRET_GETTERS_BY_KIND
 
 
 class SecretError(ValueError):
@@ -64,7 +62,9 @@ class SecretsNotFound(ExceptionGroup, SecretError):
         )
 
 
-def get_env_var_secret(source: config_secrets.EnvVarSecretSource):
+def get_env_var_secret(
+    source: config_secrets.EnvVarSecretSource,  # noqa F821 avoid cycle
+):
     if source._installation_config is not None:
         from_dotenv = source._installation_config.from_dotenv
     else:
@@ -81,10 +81,9 @@ def get_env_var_secret(source: config_secrets.EnvVarSecretSource):
         ) from exc
 
 
-_secrets_getters[config_secrets.EnvVarSecretSource.kind] = get_env_var_secret
-
-
-def get_file_path_secret(source: config_secrets.FilePathSecretSource):
+def get_file_path_secret(
+    source: config_secrets.FilePathSecretSource,  # noqa F821 avoid cycle
+):
     file_path = pathlib.Path(source.file_path)
     if not file_path.is_absolute():
         file_path = source._config_path.parent / source.file_path
@@ -99,12 +98,9 @@ def get_file_path_secret(source: config_secrets.FilePathSecretSource):
         ) from exc
 
 
-_secrets_getters[config_secrets.FilePathSecretSource.kind] = (
-    get_file_path_secret
-)
-
-
-def get_subprocess_secret(source: config_secrets.SubprocessSecretSource):
+def get_subprocess_secret(
+    source: config_secrets.SubprocessSecretSource,  # noqa F821 avoid cycle
+):
     try:
         found = subprocess.check_output(
             [source.command, *source.args],
@@ -125,47 +121,7 @@ def get_subprocess_secret(source: config_secrets.SubprocessSecretSource):
     return found.strip()
 
 
-_secrets_getters[config_secrets.SubprocessSecretSource.kind] = (
-    get_subprocess_secret
-)
-
-
-def get_random_chars_secret(source: config_secrets.RandomCharsSecretSource):
+def get_random_chars_secret(
+    source: config_secrets.RandomCharsSecretSource,  # noqa F821 avoid cycle
+):
     return os.urandom(source.n_chars).hex()
-
-
-_secrets_getters[config_secrets.RandomCharsSecretSource.kind] = (
-    get_random_chars_secret
-)
-
-
-def get_secret(secret_config: config_secrets.SecretConfig) -> str:
-    excs = []
-    sources = secret_config.sources
-    while secret_config.resolved is None and sources:
-        source, *sources = sources
-        getter = _secrets_getters[source.kind]
-        try:
-            secret_config._resolved = getter(source)
-        except SecretError as exc:
-            excs.append(exc)
-
-    if secret_config.resolved is None:
-        raise SecretSourcesFailed(secret_config.secret_name, excs)
-
-    return secret_config.resolved
-
-
-def resolve_secrets(secret_configs: list[config_secrets.SecretConfig]) -> None:
-    failed_names = []
-    excs = []
-
-    for secret_config in secret_configs:
-        try:
-            get_secret(secret_config)
-        except SecretError as exc:
-            failed_names.append(secret_config.secret_name)
-            excs.append(exc)
-
-    if failed_names:
-        raise SecretsNotFound(",".join(failed_names), excs)
