+++ tests/unit/config/test_config_installation.py
@@ -1120,7 +1120,7 @@ RaiseUnknownSecret = pytest.raises(secrets.UnknownSecret)
         ({SECRET_NAME_1: SECRET_CONFIG_1}, NoRaise),
     ],
 )
-@mock.patch("soliplex.secrets.get_secret")
+@mock.patch("soliplex.config.secrets.get_secret")
 def test_installationconfig_get_secret(gs, secret_map, expectation):
     i_config = config_installation.InstallationConfig(
         id="test-ic",
@@ -1161,7 +1161,7 @@ def test_installationconfig_get_secret(gs, secret_map, expectation):
         ),
     ],
 )
-@mock.patch("soliplex.secrets.get_secret")
+@mock.patch("soliplex.config.secrets.get_secret")
 def test_installationconfig_interpolate_secret(
     gs,
     value,
@@ -1451,7 +1451,7 @@ def test_installationconfig_interpolate_environment(
         ),
     ],
 )
-@mock.patch("soliplex.secrets.get_secret")
+@mock.patch("soliplex.config.secrets.get_secret")
 def test_installationconfig_interpolate(
     gs,
     value,
